integer revolution
- polygons can now be defined using integer arrays that get converted to vectors internally
- everything is integer

+++ src/engine/shapes/Vector2.java
@@ -1,9 +1,9 @@
 package engine.shapes;
 
 public class Vector2 {
-    double x, y;
+    int x, y;
 
-    public Vector2(double x, double y) {
+    public Vector2(int x, int y) {
         this.x = x;
         this.y = y;
     }
