+++ src/engine/shapes/Vector3.java
@@ -1,15 +1,15 @@
 package engine.shapes;
 
 public class Vector3 {
-  double x, y, z;
+  int x, y, z;
 
-    public Vector3(double x, double y, double z) {
+    public Vector3(int x, int y, int z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
     
-    public Vector2 getProjected(double focalLength) {
+    public Vector2 getProjected(int focalLength) {
       return new Vector2((focalLength*this.x)/(focalLength+this.z), (focalLength*this.y)/(focalLength+this.z));
     }
